update from August 15th

+++ src/main/java/org/liftoff/DigitalRecipeManager/DigitalRecipeManager/controllers/AuthenticationController.java
@@ -1,7 +1,7 @@
 package org.liftoff.DigitalRecipeManager.DigitalRecipeManager.controllers;
 
+import org.liftoff.DigitalRecipeManager.DigitalRecipeManager.data.UserRepository;
 import org.liftoff.DigitalRecipeManager.DigitalRecipeManager.models.User;
-import org.liftoff.DigitalRecipeManager.DigitalRecipeManager.models.data.UserRepository;
 import org.liftoff.DigitalRecipeManager.DigitalRecipeManager.models.dto.LoginFormDTO;
 import org.liftoff.DigitalRecipeManager.DigitalRecipeManager.models.dto.RegisterFormDTO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,7 +31,7 @@ public class AuthenticationController {
             return null;
         }
 
-        Optional<User> user = userRepository.findById(Long.valueOf(userId));
+        Optional<User> user = userRepository.findById(Math.toIntExact(Long.valueOf(userId)));
 
         if (user.isEmpty()) {
             return null;
